% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deidentify.R
\name{deidentify}
\alias{deidentify}
\title{Deidentify a data frame.}
\usage{
deidentify(data, ..., key = "id", drop = TRUE)
}
\arguments{
\item{data}{A data frame (or tibble).}

\item{...}{A list of the columns in `data` that contain personally
identifying information, from which the unique IDs will be generated.}

\item{key}{The name of the column to create containing unique IDs, "id" by
default.}

\item{drop}{A logical value, TRUE by default, indicating whether to remove
the personally identifying columns after the IDs are created.}
}
\description{
`deidentify()` will generate a unique ID from personally identifying
information. Because the IDs are generated with the SHA-256 algorithm, they
are a) very unlikely to be the same for people with different identifying
information, and b) nearly impossible to recover the identifying information
from.
}
\details{
This function uses non-standard evaluation for column names in `data`, so
there's no need to surround them with quotation marks.
}
